pyproject.toml
[project]
name = "agentic-document-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "kira2406",email = "kush.p030.24@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "langchain (>=1.0.2,<2.0.0)",
    "langgraph (>=1.0.1,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "tavily-python (>=0.7.12,<0.8.0)",
    "chromadb (>=1.2.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "langchain-chroma (>=1.0.0,<2.0.0)",
    "langchain-openai (>=1.0.1,<2.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
